Implement zip archive creation for downloaded fonts
Added a new functionality to generate a unique output directory for each run, download the fonts into this unique directory and create a zip archive of all downloaded fonts. This archive is stored in a temporary location. Code cleanup functions were added to remove the directory and archive upon completion.