Add search input event listeners
Two event listeners have been added in main.js. The first listener handles submit events on the search input, performs URL encoding on the input value and redirects to a download page. The second listener manages focus in events, reads text from the clipboard and if it starts with a specific string, the clipboard content is used as input value.