Display error message from URL
The code has been updated to extract an error message from the URL query string whenever an "error" parameter is present. This error message is then displayed to the user and the URL is cleaned up by removing the query string to prevent repeat display of the same error message.