Add robots.txt file
A new robots.txt file has been added to the public directory. This new file contains rules that instruct bots on how to crawl and index the website's pages, specifically disallowing access to /assets/ and /download for Googlebot, while allowing full access for all other user-agents. In addition, it points all bots to the sitemap.